+++ examples/date-picker/example-1.js
@@ -1,7 +1,0 @@
-import Datepicker from 'design-system-react/components/date-picker';
-
-<Datepicker
-	onDateChange = {function(date){
-		console.log('>>> onDateChange ', date);
-	}}
-/>