+++ examples/forms/checkbox/example-1.js
@@ -1,5 +1,0 @@
-import Checkbox from '../../../components/forms/checkbox';
-
-<Checkbox
-	assistiveText="Checkbox Example"
-	label="Checkbox Label" />