+++ examples/forms/input/example-1.js
@@ -1,7 +1,0 @@
-import Input from '../../../components/forms/input';
-
-<Input
-	id="unique-id-1"
-	label="Input Label"
-	placeholder="Placeholder Text"
-/>