+++ examples/forms/input/example-4.js
@@ -1,9 +1,0 @@
-import Input from '../../../components/forms/input';
-
-<Input
-	id="unique-id-4"
-	label="Input Label"
-	required
-	errorText="Error Message"
-	placeholder="Placeholder Text"
-/>