+++ examples/popover-tooltip/example-2.js
@@ -1,9 +1,0 @@
-import PopoverTooltip from 'design-system-react/components/popover-tooltip';
-import Button from 'design-system-react/components/button';
-
-<PopoverTooltip
-	align="right"
-	content={<span>Tooltip with right alignment</span>}
-	key="tooltipDemo">
-		<Button variant="brand" label="Hover to Open" />
-</PopoverTooltip>