+++ RELEASENOTES.md
@@ -25,6 +25,27 @@ These are changes that have backwards-compatible solutions present and that comp
 
 ---
 
+## Release 0.6.0
+
+**MINOR CHANGES**
+
+- Move `id` on GlobalNav Link to list item instead of `a`
+- In Button, `aria-*` props now use hyphen instead of camelCase. Use `aria-expanded` and not `ariaExpanded`.
+
+## Release 0.5.4
+
+**MINOR CHANGES**
+
+- Add icon/figure support to Page header (Object home variant)
+- Correct dropdown trigger styling within a button group (e.g.- more button)
+- Fix bug caused by DST and hours calculation in Datepicker.
+
+## Release 0.5.3
+
+**DOCUMENTATION**
+
+- Moves stories into examples folder, doc site examples are now aligned with SLDS website. *No production code changes in this release*.
+
 ## Release 0.5.2
 
 **MINOR CHANGES**